spotify-playlist-generator/backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
spotipy==2.19.0
python-jose==3.3.0
python-dotenv==0.19.0 